apps/api/requirements.txt
# %% Framework %%
Django>=5.1,<5.2
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.5,<6
django-cors-headers>=4.6,<5

# %% Database %%
dj-database-url>=2.3,<3
psycopg2-binary>=2.9,<3

# %% Async tasks %%
celery>=5.4,<6
django-celery-beat>=2.7,<3
redis>=5.0,<7

# %% HTTP / Payments %%
requests>=2.31,<3

# %% Data / Import-Export %%
pandas>=2.2,<3
openpyxl>=3.1,<4

# %% Forecast (opcional — fallback a WMA si no está) %%
statsmodels>=0.14,<1

# %% Server %%
gunicorn>=22,<24
whitenoise>=6.7,<7

# %% Config %%
python-dotenv>=1.0,<2
